Replace is_strictly_increasing method on Dimension with SeekBias
In a subsequent commits I want to associate anchors with a right or left
bias, which will affect whether subsequent insertions at the anchor's
position push the anchor to the right. To respect this bias, it will
help to be able to seek in the insertion split trees with an explicitly
specified bias indicating whether I want the leftmost or rightmost item
that inclusively contains the target position.

Also, for some kinds of summarized properties such as the
max_fragment_id, the concept of "strictly increasing" doesn't really
make sense on the dimension. The identifiers are all unique and
increasing, but we still want to seek to the leftmost leaf with that max
identifier.

/cc @as-cii